requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
python-dotenv>=1.0.1
Jinja2>=3.1.0
openai>=1.50.0
google-genai
httpx>=0.27.0
solapi>=0.1.0
